refac: clean debate call handler
