+++ book-talk-fe/src/hooks/useDebateWebSocket.tsx
@@ -1,11 +1,25 @@
 import {useCallback, useEffect, useMemo, useRef, useState} from 'react';
+import {useQueryClient} from "@tanstack/react-query";
 import {DebateWebSocketClient, type WebSocketHandlers} from "../apis/websocket";
+import {findOneDebateQueryOptions} from "../apis/debate";
+import type {DebateRoundInfo} from "../apis/websocket/client.ts";
+
+type RoundType = 'PREPARATION' | 'PRESENTATION' | 'FREE';
+
+interface UseDebateWebSocketOptions {
+    onRoundStartBackdrop?: (roundType: RoundType) => void;
+}
 
 /**
  * 토론방의 실시간 WebSocket 연결을 관리하는 훅
  * WebSocket을 통해 토론방의 모든 실시간 이벤트를 처리합니다.
  */
-export const useDebateWebSocket = (debateId: string | null, handlers: WebSocketHandlers) => {
+export const useDebateWebSocket = (
+    debateId: string | null,
+    handlers: WebSocketHandlers,
+    options?: UseDebateWebSocketOptions
+) => {
+    const queryClient = useQueryClient();
     const [onlineAccountIds, setOnlineAccountIds] = useState<Set<string>>(new Set());
     const [isConnected, setIsConnected] = useState<boolean>(false);
     const [raisedHands, setRaisedHands] = useState<Array<{
@@ -16,7 +30,28 @@ export const useDebateWebSocket = (debateId: string | null, handlers: WebSocketH
     const wsClientRef = useRef<DebateWebSocketClient | null>(null);
     const heartbeatIntervalRef = useRef<number | null>(null);
 
-    // 핸들러와 내부 상태 관리를 결합 (메모이제이션)
+    /** 발언자 업데이트 콜백 */
+    const handleSpeakerUpdate = useCallback((speakerInfo: unknown) => {
+        console.log('Speaker updated via WebSocket:', speakerInfo);
+        if (debateId) {
+            void queryClient.invalidateQueries({queryKey: findOneDebateQueryOptions().queryKey});
+        }
+        handlers.onSpeakerUpdate?.(speakerInfo);
+    }, [debateId, handlers, queryClient]);
+
+    /** 토론 업데이트 콜백 */
+    const handleDebateRoundUpdate = useCallback((roundInfo: DebateRoundInfo) => {
+        console.log('Debate round updated via WebSocket:', roundInfo);
+        const roundType = roundInfo.round.type as RoundType;
+        if (roundType === "PRESENTATION" || roundType === "FREE") {
+            options?.onRoundStartBackdrop?.(roundType);
+        }
+        if (debateId) {
+            void queryClient.invalidateQueries({queryKey: findOneDebateQueryOptions().queryKey});
+        }
+        handlers.onDebateRoundUpdate?.(roundInfo);
+    }, [debateId, handlers, options, queryClient]);
+
     const combinedHandlers = useMemo(() => ({
         onPresenceUpdate: (onlineIds: Set<string>) => {
             console.log('Received online account IDs:', onlineIds);
@@ -28,28 +63,24 @@ export const useDebateWebSocket = (debateId: string | null, handlers: WebSocketH
             setIsConnected(connected);
             handlers.onConnectionStatus?.(connected);
         },
-        onHandRaiseUpdate: (hands: Array<{accountId: string; accountName: string; raisedAt: number}>) => {
+        onHandRaiseUpdate: (hands: Array<{ accountId: string; accountName: string; raisedAt: number }>) => {
             console.log('Received raised hands update:', hands);
             setRaisedHands(hands);
             handlers.onHandRaiseUpdate?.(hands);
         },
-        onSpeakerUpdate: handlers.onSpeakerUpdate,
-        onDebateRoundUpdate: handlers.onDebateRoundUpdate,
+        onSpeakerUpdate: handleSpeakerUpdate,
+        onDebateRoundUpdate: handleDebateRoundUpdate,
         onVoiceSignaling: handlers.onVoiceSignaling
-    }), [handlers]);
+    }), [handlers, handleSpeakerUpdate, handleDebateRoundUpdate]);
 
-    // WebSocket 연결 및 관리
+    /** WebSocket 연결 및 관리 */
     useEffect(() => {
         if (!debateId) return;
-
-        // DebateWebSocketClient 인스턴스 생성
         const wsClient = new DebateWebSocketClient();
         wsClientRef.current = wsClient;
 
-        // 연결 시작
         wsClient.connect(debateId, combinedHandlers);
 
-        // 정리 함수
         return () => {
             if (heartbeatIntervalRef.current) {
                 clearInterval(heartbeatIntervalRef.current);
@@ -63,7 +94,7 @@ export const useDebateWebSocket = (debateId: string | null, handlers: WebSocketH
         };
     }, [debateId, combinedHandlers]);
 
-    // 하트비트 관리
+    /** WebSocket 하트비트 관리 */
     useEffect(() => {
         if (isConnected && wsClientRef.current) {
             // 30초마다 하트비트 전송
@@ -93,6 +124,7 @@ export const useDebateWebSocket = (debateId: string | null, handlers: WebSocketH
         return result;
     }, [onlineAccountIds]);
 
+    /** 손들기 기능 */
     const toggleHand = useCallback(() => {
         wsClientRef.current?.toggleHand();
     }, []);
